+++ CHANGES.rst
@@ -5,6 +5,9 @@
 4.7.3 (unreleased)
 ==================
 
+- Fix ``Tree.__setstate__`` to no longer accept children besides
+  tree or bucket types to prevent crashes. See `PR 143
+  <https://github.com/zopefoundation/BTrees/pull/143>`_ for details.
 - BTrees, TreeSet, Set and Buckets implements the ``__and__``,
   ``__or__`` and ``__sub__`` as shortcuts for
   :meth:`~BTrees.Interfaces.IMerge.intersection`,