+++ CHANGES.rst
@@ -10,9 +10,9 @@
   <https://github.com/zopefoundation/BTrees/pull/143>`_ for details.
 - BTrees, TreeSet, Set and Buckets implements the ``__and__``,
   ``__or__`` and ``__sub__`` as shortcuts for
-  :meth:`~BTrees.Interfaces.IMerge.intersection`,
-  :meth:`~BTrees.Interfaces.IMerge.union` and
-  :meth:`~BTrees.Interfaces.IMerge.difference`.
+  ``~BTrees.Interfaces.IMerge.intersection``,
+  ``~BTrees.Interfaces.IMerge.union`` and
+  ``~BTrees.Interfaces.IMerge.difference``.
 
 
 4.7.2 (2020-04-07)