+++ CHANGES.rst
@@ -5,6 +5,9 @@
 4.7.3 (unreleased)
 ==================
 
+- Fix persistency bug in the Python version
+  (`#118 <https://github.com/zopefoundation/BTrees/issues/118>`_).
+
 - Fix ``Tree.__setstate__`` to no longer accept children besides
   tree or bucket types to prevent crashes. See `PR 143
   <https://github.com/zopefoundation/BTrees/pull/143>`_ for details.